Phân Category theo dạng Parent-Child

+++ WebsiteBook/Repositories/EFCategoryRepository.cs
@@ -20,19 +20,25 @@ namespace FashionShopDemo.Repositories
         public async Task DeleteAsync(int id)
         {
             var category = await _context.Categories.FindAsync(id);
-            _context.Categories.Remove(category);
-            await _context.SaveChangesAsync();
+            if (category != null)
+            {
+                _context.Categories.Remove(category);
+                await _context.SaveChangesAsync();
+            }
         }
 
         public async Task<IEnumerable<Category>> GetAllAsync()
         {
             return await _context.Categories
-            .ToListAsync();
+                .Include(c => c.SubCategories) // danh mục con
+                .ToListAsync();
         }
 
         public async Task<Category> GetByIdAsync(int id)
         {
-            return await _context.Categories.FirstOrDefaultAsync(p => p.Id == id);
+            return await _context.Categories
+                .Include(c => c.SubCategories) // danh mục con
+                .FirstOrDefaultAsync(c => c.Id == id);
         }
 
         public async Task UpdateAsync(Category category)
@@ -40,5 +46,23 @@ namespace FashionShopDemo.Repositories
             _context.Categories.Update(category);
             await _context.SaveChangesAsync();
         }
+        public async Task<List<Category>> GetCategoryTreeAsync()
+        {
+            var categories = await _context.Categories.ToListAsync();
+            return categories
+                .Where(c => c.ParentId == null) //  danh mục gốc
+                .Select(c => BuildCategoryTree(c, categories)) // cây danh mục
+                .ToList();
+        }
+
+        private Category BuildCategoryTree(Category category, List<Category> allCategories)
+        {
+            category.SubCategories = allCategories
+                .Where(c => c.ParentId == category.Id)
+                .Select(c => BuildCategoryTree(c, allCategories))
+                .ToList();
+            return category;
+        }
+
     }
 }
